backend/requirements.txt
fastapi
uvicorn
pydantic[email]
python-dotenv
aiosmtplib
